Split rewards_program.rs
