+++ programs/native/rewards_api/src/rewards_instruction.rs
@@ -1,4 +1,4 @@
-use crate::rewards_program;
+use crate::id;
 use serde_derive::{Deserialize, Serialize};
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::transaction_builder::BuilderInstruction;
@@ -11,7 +11,7 @@ pub enum RewardsInstruction {
 impl RewardsInstruction {
     pub fn new_redeem_vote_credits(vote_id: Pubkey, rewards_id: Pubkey) -> BuilderInstruction {
         BuilderInstruction::new(
-            rewards_program::id(),
+            id(),
             &RewardsInstruction::RedeemVoteCredits,
             vec![(vote_id, true), (rewards_id, false)],
         )
